requirements.txt
transformers==4.45.1
accelerate
tiktoken
einops
transformers_stream_generator
scipy
torchvision
pillow
tensorboard
matplotlib
qwen-vl-utils